stm32: fix flash_ec stm32 serial EC programming mode
Proper sequence of events when enabling STM32 monitor mode is as
follows:

  - bootp0 pin level is set to 'high'

  - chip is reset

  - the programmer starts sending sync bytes over the UART

This patch implements this sequence in flash_stm32(). In cases when
'cold_reset' control is not available this function will fail, as this
means that the STM32 chip can not be reset.

When programming over Cr50, baud rate is set to 57600, as it does not
support the full speed, and the command to enable bit bang mode is
sent to servod.

BRANCH=cr50, cr50-mp
BUG=b:62539385
TEST=successfully ran to completion (100 iterations) the following
     script when programming Scarlet when servo is connected over
     Cr50:

    count=0;
    while ./util/flash_ec --board scarlet --image <image path>; do
      count=$(( count + 1 ))
      echo "succeeded $count times"
      if [ $count == 100 ]; then
        break
      fi
      sleep 3
    done

Change-Id: Ib3242a92c94f64214ffd1f290b97c2660ce21360
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>