DragonEgg: Enable/disable battery learn mode based on AC
This CL is a software workaround for an issue with the bq25710
charger. There is an issue with the converter control loop not being
biased correctly when it starts switching. This leads to a reverse
current (from battery to charger). To avoid this issue, the switching
converter is turned off when AC is not present, then only enalbed 200
msec after AC presence is detected.

BUG=b:112372451
BRANCH=none
TEST=

Change-Id: I4d3a975e3c0e24cdf7f13fdab69da32ccd70a428
Signed-off-by: Scott Collyer <scollyer@google.com>