Nami: Set initial fan mode to TACHO_FAN_RPM
Currently after sysjump, the fan mode is reset to TACHO_FAN_DUTY. This
causes the thermal fan control not to work.

This change sets the fan mode to TACHO_FAN_RPM in fan_config. This is
NPCX specific change.

In the master, we should fix this differently: save and restore the fan
mode before and after a jump.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:112293333
BRANCH=nami
TEST=Boot Akali. Verify the fan spins after sysjump on dev screen.

Change-Id: Ief5a0b4086ced63ffbae34cec20564658e2acfd8