Update patch set 2
Patch Set 2:

In the common/charge_manager.c

#if defined(CONFIG_USB_PD_VBUS_MEASURE_CHARGER)
                        r->meas.voltage_now = charger_get_vbus_voltage(port);
#elif defined(CONFIG_USB_PD_VBUS_MEASURE_ADC_EACH_PORT)
                        r->meas.voltage_now =
                                adc_read_channel(board_get_vbus_adc(port));
#elif defined(CONFIG_USB_PD_VBUS_MEASURE_NOT_PRESENT)
                        /* No VBUS ADC channel - voltage is unknown */
                        r->meas.voltage_now = 0;
#else
                        /* There is a single ADC that measures joint Vbus */
                        r->meas.voltage_now = adc_read_channel(ADC_VBUS);
#endif

If the CONFIG_USB_PD_VBUS_MEASURE_NOT_PRESENT not defined, it tries to read the -1 adc channel. A safer way is probably to move the "NO_PRESENT" to the "else" clause. So it gets a safe "0" value if none of these configs is defined.

Patch-set: 2