nocturne: Update sensor active mask to include S3.
Previously, the active_mask member in motion_sensor_t was set to the S0
for all the sensors.  Although the sensors are only used in S0, this
value should reflect the actual power state of the sensor.  The config
members describe the power states when the EC intends to use the sensor.
This combination was causing the sensors to not have their calibration
data applied following a suspend/resume cycle.

This commit updates the active mask to include S3 thereby keeping the
calibration settings across a suspend resume cycle.

BUG=b:112462750, b:111815348
BRANCH=None
TEST=Flash nocturne; set calibration attributes in VPD, reboot, verify
that calibration attributes are applied.  Suspend/resume, verify that
the calibration attributes are still applied.

Change-Id: Iad632e3bb6f61b056015ca8c9d90b0f9d7c58c5e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>