Update patch set 2
Patch Set 2:

Although this CL fixes the problem at hand, it does not address the unerlying bug:
When AP calls SENSOR_RANGE, the new range is not saved. So when init path is executed, in sensor_init_done(), the default range is re-applied.
With this change, we are not calling sensor_init_done unless we are coming from S5 to S3, and luckily, the host will call SENSOR_RANGE at every reboot.

For sensors that are really powered only in S0, we still have a problem.
We could save the new range in default_range, but that will confuse dual booting systems.

We have a similar problem with SENSOR_OFFSET, even worse actually, because there is not provision for a default offset field, but only for sensors that are working only in S0 and manage their offset themselves.

We can work around the problem in the kernel by saving and sending OFFSET/RANGE at resume time, but the intent was to have kernel driver as stateless as possible.

Patch-set: 2
Reviewer: Gerrit User 1104704 <1104704@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>