chgstv2: Use lower charge threshold for boot if AC power is present
If one of the following conditions is met, EC boots AP:

1. Battery charge >= CONFIG_CHARGER_MIN_BAT_PCT_FOR_POWER_ON
2. AC power >= CONFIG_CHARGER_MIN_POWER_MW_FOR_POWER_ON
3. Battery charge >= CONFIG_CHARGER_MIN_BAT_PCT_FOR_POWER_ON_WITH_AC
   and
   AC power >= CONFIG_CHARGER_MIN_POWER_MW_FOR_POWER_ON_WITH_BATT

This reduces users' wait time significantly when a battery is drained.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:112172032
BRANCH=Nami
TEST=Verified Akali and Sona boot and perform soft sync successfully.
The power consumption is measured during boot and verified even worn
out battery can supply enough power to sustain the system together with
a 15W charger.

Change-Id: I3220b4660a1b50ba5ead8f570cb6992c0a30e390