Update patch set 1
Patch Set 1: Verified+1

This CL adds oss-fuzz support to the EC codebase. My plan is to fuzz the vendor specific commands and possibly the TPM2 stack (even though it is already fuzzed separately) because as far as I know the original fuzzer isn't aware of the semantics of TPM2 messages.

Part of this process was adding support to build c++ sources and protocol buffers as this is a requirement for libprotobuf-mutator. One of the header files fails the format checking because it is an object header and uses C++ style.

I have included a fuzzer target for Cr50 that currently only provides semantics for three pinweaver commands, so I could get feedback on what is needed to merge this prior to building out the reset of the fuzzer. I have a working oss-fuzz configuration for it as well, but I don't plan on uploading it yet.

There isn't an easy way to build this target outside of the oss-fuzz environment, but I have managed it get it to work for testing on my workstation using a dummy libFuzzingEngine and the right environmental variables.

Patch-set: 1
Reviewer: Gerrit User 1152319 <1152319@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1135824 <1135824@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1160200 <1160200@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1