Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This is a general question: Why do we need to add all this extra code to handle 2 different fuel guages? I probably missed the conversation yesterday, but this seems like a large amount code that needs to be added. Don't we have generic drivers for the fuel guages and common code to utilize?

We don't need to do this. I think there is some misunderstanding going on. The mfgacc_support enables access to the manufacturer access registers, which could be different between TI and Reneas. But, the fuel gauge registers as defined in smart_battery.h are common to both. 

These are all the same batteries that were used for Coral and were supported with the .fuel_gauge structure. The field .mfgacc_support field was added by ODM partners who thought that they needed to access the operational_status (0x54) register to check the FET status. I pointed out at the time that the lower 16 bits of this register are returned in register 0x0 and shouldn't need to access 0x54 to check the FET status. The only place that .mfgacc_support is used is to get the FET status.

Patch-set: 3
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>