common: add API to expose detachable "base" state
On some detachables, when base is attached, we know right away that the
device should transition from tablet to clamshell mode. However on other
detachables we need additional information (i.e. base position) before
we decide whether to transition in/out of tablet mode. For such
detachables let's allow them to signal a new "base attached" switch
event, so that the rest of the stack is not confused.

BUG=b:73133611
BRANCH=nocturne
TEST=Build and boot

Signed-off-by: RaviChandra Sadineni <ravisadineni@google.com>

Change-Id: I3154f356ff2981a5edb2c31cde0d255dc49e9737
orig-Change-Id: I9be3450cba52bf9f0bad8333402f68b0c7903090
orig-Reviewed-on: https://chromium-review.googlesource.com/1176801