chip/stm32: Convert usb_endpoints to C so gcc's LTO knows about it
If we keep it assembly-only, the link time optimizer gets confused and
eliminates seemingly unused functions, to then replace references to
them with the "no handler" defaults in a later step.

Similar approach as with vecttable: Implement the table in C so LTO
knows the entire story.

BUG=b:65441143
BRANCH=none
TEST=usb_ep_{rx,tx,event} and usb_iface_request look more reasonable in
disassembly on whiskers.

Change-Id: I35ccfd68cda2d0022aa464ecf622f4eef71c3398
Signed-off-by: Patrick Georgi <pgeorgi@google.com>