ppc: Add support for nx20p3481
This CL is an incremental change to the nx20p348x driver to add
support for the nx20p3481 ppc. Sink/source modes are controlled via
the switch control register instead of gpio signals. In addition, the
nx20p3481 supports Fast Role Swap (FRS).

For FRS, only the detection is supported, and it's assumed that it's
caused by the removal of an external charger, not an actual FRS event.

BUG=b:111281797
BRANCH=none
TEST=Verified on DragonEgg that port acts correctly as a sink. Have
not been able to verify source operation.

Change-Id: I2fb4200a5d9c3ce460e9b913a5b09441e458bb7e
Signed-off-by: Scott Collyer <scollyer@google.com>