fuzz: Hide conflicts with cstdlib and use clang++ for linking.
This creates a build target called libec.a by setting the visibility
of functions that conflict with cstdlib to hidden. It then links
those symbols locally into one large object file that makes up libec.a

Fuzzing targets are linked against libec.a so that they can invoke ec
functionality while depending on outside libraries that need cstdlib.

The fuzzing targets are now linked using clang++, so that c++ libraries
and objects can be used as part of the fuzzers.

CQ-DEPEND=CL:1180179
BRANCH=none
BUG=none
TEST=make -j buildfuzztests &&
  ./build/host/host_command_fuzz/host_command_fuzz.exe

Change-Id: Ifdfdc6a51c6ef23b4e192b013ca993bf48a4411b
Signed-off-by: Allen Webb <allenwebb@google.com>