nocturne: Increase priority of sensor hooks.
For nocturne, the sensors are completely unpowered in S5.  Therefore,
the motionsense task needs to be notified quickly of chipset transitions
to avoid attempting to access the unpowered sensors.  When it did this,
it would lead to a watchdog reset while trying to unwedge the i2c bus.

This commit simply increases the hook priority for the sensor related
hooks and motion sense hooks.  This way the motion sense task can be
notified of the chipset transitions quicker.

BUG=b:111683988
BRANCH=None
TEST=Flash nocturne; Boot from S5 to S0 and down to S5 several times.
Verify that no EC reset occurs.
TEST=Also verify that no i2c unwedge prints are seen for the sensor bus.

Change-Id: Idb860b2e62b04d0e8d3bdbf55e6c6b7a439b8b54
Signed-off-by: Aseda Aboagye <aaboagye@google.com>