ec3po: quit console & interpreter loop when parent process changes.
Right now, when the parent process dies ungracefully - say kill -9 -
then the interpreter, and console processes remain active.
This leads to bugs in the servod implementation from holding on to
sockets, to reinitialization issues of a new instance on the same servod
device.
This change quits the loops inside console & interpreter as soon as the
parent pid changes (i.e. the parent dies).

BRANCH=None
BUG=chromium:841121
TEST=sudo servod -b soraka
ps aux | grep servod
>xxxxx servod
>xxxxy servod
>xxxyx servod
>xxxaa grep servod
sudo kill -9 xxxxx
ps aux | grep servod
>xxxab grep servod

Before this, just kill -9 on the main thread did not take the children
with it.

Change-Id: I305bdbcc7c6d0ffc846b730c897e364b50490214
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>