bmi160: End IRQ handler if sensor isn't powered.
It's possible for the interrupt to be triggered and then when the bottom
half of the interrupt handler gets a chance to run, the sensor isn't
powered anymore.  This commit simply has the loop terminate early if the
board indicates that the bus the sensor is on is no longer powered.

BUG=b:111683988
BRANCH=nocturne
TEST=Verify if IRQ handler is called when sensor is not powered, the
irq_handler loop terminates.

Change-Id: Iaf395902bb4b46a5b6d750c99767c4c36b1e7a99
Signed-off-by: Aseda Aboagye <aaboagye@google.com>