usb-pd: disconnect USB SS MUX when S5 or lower
Especially for SS MUX that have redrivers in them, we should disable the
MUX while the chipset is in suspend because the data line will not be
used. This give decent power savings for redriver MUXs (e.g. PS8751)

BRANCH=none
BUG=b:112136208
TEST=On Phaser the 3300_pd_a drops from 92mW to 32 mW when the charger
is plugged into C1 and the SoC is in S5. The rail also says at 32mW after
removing and plugging the power back in while the SoC is in S5. Also
ensured that power is low upon first insertion and AP does not come on
automatically.

Change-Id: I0601fbb506ad3eff902cf6562a6408292ef70e3a
Signed-off-by: Jett Rink <jettrink@chromium.org>