yorp: increasing TCPC i2c speed causes instability
When attempting to increase our i2c speeds on the TCPC buses from 100k
to 400k, we started seeing stability issues on the USB-C ports.  With
the speed increase, we were no longer giving enough time for VBUS to
rise before attempting to check the status of it.  This change adds a 2
ms delay before checking the VBUS status in the PPC driver.

BRANCH=None
BUG=b:78554726
TEST=Loaded onto yorp proto 2 and confirmed that charging and display on
ports 0 and 1 work with 400k speed

Change-Id: I73f4e175184c9d4f791ea5b84dc91ba6821ce8a7
Signed-off-by: Diana Z <dzigterman@chromium.org>