rdd: use rdd_interrupt for rdd_connect/disconnect
Occasionally cc1 and cc2 might be in a valid debug range for a short
time even if a debug accesssory isn't connected. The rdd hardware will
debounce the signals long enough to make sure they're stable before
triggering the Rdd interrupt. This change moves all rdd
connect/disconnect logic to the interrupt instead of sampling the
voltages in detect_rdd. In detect_rdd the samples of the cc lines
falsely triggered rdd connect whenever the cc voltages happened to be in
the valid debug range. We weren't waiting for Rdd to declare the voltage
stable before running rdd connect. Moving the logic from rdd_detect to
rdd_interrupt will ensure the signals are actually stable before we
change the rdd state.

BUG=b:113064204
BRANCH=cr50
TEST=manual
	connect suzyq
	reboot cr50
	verify rdd is connected

	disconnect suzyq
	verify rdd is disconnected

	boot nami
	leave it up for a while
	make sure there aren't any false connects

	reconnect suzyq
	verify rdd is connected

	disconnect suzyq
	verify rdd is disconnected

	reboot cr50
	verify rdd is disconnected

	run firmware_Cr50CCDServoCap and firmware_Cr50DeviceState

Change-Id: I3eeaa39f79d2ccc3e997da291d8aae1f12103dc8
Signed-off-by: Mary Ruthven <mruthven@google.com>