rdd: only use rdd_interrupt for rdd_connect
Occasionally cc1 and cc2 might be in a valid debug range for a short
time even if a debug accesssory isn't connected. The rdd hardware will
normally handle making sure they stay in that range for a long enough
time to determine a device is actually connected. rdd_detect was just
sampling the signals once, so it was incorrectly connecting rdd every so
often even though a rdd accessory wasn't connected. This changes rdd to
only call rdd_connect from the interrupt to make sure the connect is
valid.

BUG=b:113064204
BRANCH=cr50
TEST=manual
	connect suzyq
	reboot cr50
	verify rdd is connected

	disconnect suzyq
	verify rdd is disconnected

	boot nami
	leave it up for a while
	make sure there aren't any false connects

	reconnect suzyq
	verify rdd is connected

	disconnect suzyq
	verify rdd is disconnected

	reboot cr50
	verify rdd is disconnected

	run firmware_Cr50CCDServoCap and firmware_Cr50DeviceState

Change-Id: I3eeaa39f79d2ccc3e997da291d8aae1f12103dc8
Signed-off-by: Mary Ruthven <mruthven@google.com>