base_detect: Expose console command to force state.
In an effort to test wake sources on any given platform, this
CL exposes console command to set the base state. This console
command can then be invoked by autottests from the uart interface.
On LUX, soraka and cheza :
     1. For force connect and disconnect :
           a. we disable interrupts and force the base state.
     2. For reset:
           b. We schedule the differed task immediately and
	      enable interrupts.
On nocturne and zoombini we don't depend on interrupt to identify
the base state. Instead we have a periodic task to check if base
is attached. On forcing a state, we stop rescheduling the
periodic task after forcing the set base state. On reset, we
start scheduling the task to detect base periodically again.

Signed-off-by: RaviChandra Sadineni <ravisadineni@google.com>
BRANCH=poppy,nocturne
BUG=chromium:820668
TEST=Tested on lux, soraka and nocturne
             basestate  a : attaches the lid, reflected in ui.
             basestate  d : detaches the lid, reflected in ui.
             basestate  r : resets to the correct state.
     Wakes the device up on lux and nocturne and soraka.

Change-Id: Iab698e103a50b8d22bf216a6f816998cb158e38a