Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> Does this actually happen? Can we fix the charger state machine instead?

"Does this actually happen" and "Can we" covers a lot of ground.  "Does this happen" depends on just exactly how the EC was configured.  It is not practical to audit every combination that exists due to the many options available, in hardware, hardware configuration, and config.h options.

Yes, the charger state machine can be fixed.  The descendant CL to this one describes what must be done.  But there is a maintainability problem with this approach and I would rather not take it.  The application of the current limit is a complex process that is spread across several call sites, and based on different needs.  By saturating the current limit to respect just the needs of the hardware in just the hardware's driver (as done in this CL, not the child), we can cleanly separate out that particular concern from the remainder of the code.

> We are not normally that defensive in the EC (because of code space...)

This is a case where we are defensive only some of the time.  There are some paths that respect the driver's advertised limits, and some paths that do not.

>, so I'm worried about sprinkling seemingly well-intended checks in the code.
> 
> Please also look at code size impact (make savesizes before this change, and make newsizes after).

This is a fairly small change in either case.  Using GCC 4.9 (which the EC has reverted back to for now), the delta with this CL alone is +51 bytes avg, +144 bytes peak.  With the child CL alone, the delta is +59 bytes avg, +192 bytes peak.  Fixing the problems in chgstv2 costs slightly more space, since there are more places to fix, but they are practically equal.

The important problem is that doing it in the callers makes ensuring correctness harder, as evidenced by the number of holes which are present.

Patch-set: 2
Reviewer: Gerrit User 1284593 <1284593@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>