servo_micro: Allow setting the baud rate for usart
We set the baud rate in increments of 100 baud, to avoid
overflowing the 16-bit wValue integer (921600 is the highest we
are likely to use).

Also, increment the buffer size for USART3 to 1024 bytes. That
helps a bit to avoid losing characters, but it appears that the
problem is that we are hitting the limit of what bulk endpoints
can provide in USB FS.

BRANCH=servo
BUG=chromium:876651
TEST=baud usart2/3/4 115200 in servo_micro console
TEST=dut-control cpu_uart_baudrate:921600
     seq 1 1000 shows numbers 1 to 226 before buffer overflows

Change-Id: Ifca266189f93def493f207dd29d2cceca4d8d68f
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>