Fleex: Enable dynamic motion sensor count and change disable the keyboard for convertible
Get the motion sensor count by SKU
motion_sensor_count on clamshells is 0.
motion_sensor_count on convertibles is 2.

lid_angle_peripheral_enable() function disable
the keyboard for convertible systems.

BUG=b:112286331
BRANCH=none
TEST=Manual
change SKUID to 0x21 for convertible and verified the motion sensor
count and that the motion senors were initialized in the EC console log.
change SKUID to 0x01 for clamshell and the motion senors were
not initialized in the EC console log.

Change-Id: Ic9d6c910c9e56479414d30f1e2c53d73515842af
Signed-off-by: elthanhuang <elthan_huang@compal.corp-partner.google.com>