nocturne: Use USBC alert signal for TCPC alert status
Because the TCPC and PPC share a common interrupt signal, the TCPC
alert register must be read to know when to process alerts for the
TCPC. However, the interrupt signal can be used to know which port to
check and prevent interrupts from one port causing the TCPC on another
port from being pulled out of low power mode unecessarily.

This CL adds a check of the USBC interrupt line level to know which
TCPC to check.

BUG=b:112039224
BRANCH=nocturne
TEST=Verifed that connecting to 1 of the ports no longer causes the
other port to exit/enter low power mode.

Change-Id: I70384a1fc8f1cc03ad81e694a4246e0cb284375f
Signed-off-by: Scott Collyer <scollyer@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1196044
Commit-Ready: Scott Collyer <scollyer@chromium.org>
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>