gsctool: refactor USB interface
Communications with Cr50 exposed USB endpoints could be needed by
other utilities, in particular, ./util/iteflash when it is extended to
operate over Cr50.

This patch moves USB interface functions into a separate file in the
./util directory and makes USB endpoint coordinates run time
variables, so that the user of the interface can connect to various
endpoints.

Some refactoring is required to allow using the generic USB transfer
function.

BRANCH=none
BUG=b:75976718
TEST=verified that gsctool still operates properly - updated a Cr50,
     read Cr50 version number, etc.

Change-Id: I3d77a93932f5395fff0f5823f0dd79e1d1d670c8
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>