Update patch set 4
Patch Set 4:

Thought a bit about this. I think you need this because there is an abstraction mismatch between the protobuf representation and the serialized byte representation that is consumed by pinweaver code. In other words, the protobuf representation doesn't let you express all inputs to the fuzzer, but only a subset that is slightly larger than the set of valid inputs, but doesn't include all invalid inputs.

That suggests what you really want the fuzzer to do is represent inputs not as serialized proto messages, but as byte strings that would be received by pinweaver. This prevents you from making use of libprotobuf-mutator though... unless you decode the byte string into a proto message, then run the proto mutator, then encode again. That'd require decoding logic as well though.

Note that for the case where an input that has been degraded into bytes and allows to find the fuzzer some interesting branches in the byte representation, you'll never be able to move back to the proto representation (which might enable further exploration).

Bottom line is that it feels like protobuf mutation works best for code that actually consumes protobuf inputs, and in this case you're kind of forcing it into a use case that it doesn't support so well. Maybe it's good enough though!

Anyhow, that's just random thoughts, I defer to Vitaly to provide qualified feedback on whether this is good use of the fuzzing infrastructure or not. Happy to stamp whatever conclusion the two of you reach.

Patch-set: 4
Reviewer: Gerrit User 1001348 <1001348@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1135824 <1135824@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>