rammus: Fix the dut keeps reboot due to the EC assert.
The dut keeps reboot due to the EC Assert. The problem is caused by
the changing of I2C_PORT_TCPC1 to NPCX_I2C_PORT0_0 (cl:1189663) and
the NPCX_I2C_PORT0_1 is removed from the i2c_ports[]. As a result,
when someone uses NPCX_I2C_PORT0_1, the assert popped out.
This patch changes the 'name' and 'port' of i2c_ports[] to avoid this
problem.

BUG=b:113715116
BRANCH=master
TEST=dut boot successfully

Change-Id: I90f7a1dcdec3ff2719f34786e9e57411b703562b
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>