Update patch set 5
Patch Set 5:

Are most of the grunt family boards using a PWM for their LEDs?  It could be more worthwhile to make a generic PWM LED state machine if that's the case.  For octopus, none of our boards are currently using one so the state code is written just for on/off behavior

Patch-set: 5
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>